Revert "chore(release): move release-notes tests into scripts/test/"
This reverts commit 1f3b2089.
